refactor(fem): split results/artefacts.py into the artefacts package
The flat 3,248-line module becomes a package split by artefact kind:
formats (binary layout constants and version pins), specs, protocol,
history, stream_adapter, beam_solids, mesh, fields, manifest, readers,
bake and posters. The package __init__ re-exports every public and
underscore-prefixed name so existing importers and tests keep working
unchanged; the stream-reader registry dict is shared, not copied.

No behaviour change; no new module-import-time dependencies.
